# Use Stream.CopyTo{Async} in NonCryptographicHashAlgorithm.Append{Async}